feat: add api CRUD and download file Banner
