--- src/main/java/org/ait/library/blob/config/GlobalConfig.java
@@ -1,11 +1,0 @@
-package org.ait.library.blob.config;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.ait.library.blob.shared.enums.PlatformEnum;
-
-@Setter
-@Getter
-public class GlobalConfig {
-  private PlatformEnum platform;
-}
