--- src/main/java/org/ait/library/blob/modules/selector/service/StorageSelectorService.java
@@ -1,8 +1,0 @@
-package org.ait.library.blob.modules.selector.service;
-
-
-import org.ait.library.blob.shared.serviceskelenton.EngineStorageService;
-
-public interface StorageSelectorService {
-  EngineStorageService getEngineStorageService();
-}
