--- src/main/java/org/ait/library/blob/shared/exception/CannotUploadFileException.java
@@ -1,11 +1,0 @@
-package org.ait.library.blob.shared.exception;
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class CannotUploadFileException extends RuntimeException {
-  public CannotUploadFileException(String message) {
-    super(message);
-    log.error(String.format("Cannot Upload File: %s", message));
-  }
-}
