--- src/main/java/org/ait/library/blob/shared/exception/PlatformNotFoundException.java
@@ -1,7 +1,0 @@
-package org.ait.library.blob.shared.exception;
-
-public class PlatformNotFoundException extends RuntimeException {
-  public PlatformNotFoundException() {
-    super();
-  }
-}
