--- src/main/java/org/ait/library/blob/shared/exception/StorageFileNotFoundException.java
@@ -1,7 +1,0 @@
-package org.ait.library.blob.shared.exception;
-
-public class StorageFileNotFoundException extends RuntimeException {
-  public StorageFileNotFoundException() {
-    super();
-  }
-}
