+++ src/main/java/org/ait/project/guideline/example/blob/config/GlobalConfig.java
@@ -1,0 +1,11 @@
+package org.ait.project.guideline.example.blob.config;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.ait.project.guideline.example.blob.shared.enums.PlatformEnum;
+
+@Setter
+@Getter
+public class GlobalConfig {
+  private PlatformEnum platform;
+}
