+++ src/main/java/org/ait/project/guideline/example/blob/modules/selector/service/StorageSelectorService.java
@@ -1,0 +1,8 @@
+package org.ait.project.guideline.example.blob.modules.selector.service;
+
+
+import org.ait.project.guideline.example.blob.shared.serviceskelenton.EngineStorageService;
+
+public interface StorageSelectorService {
+  EngineStorageService getEngineStorageService();
+}
