+++ src/main/java/org/ait/project/guideline/example/blob/shared/exception/CannotDownloadFileException.java
@@ -1,0 +1,11 @@
+package org.ait.project.guideline.example.blob.shared.exception;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class CannotDownloadFileException extends RuntimeException {
+  public CannotDownloadFileException(String message) {
+    super(message);
+    log.error(String.format("Cannot Download File: %s", message));
+  }
+}
