+++ src/main/java/org/ait/project/guideline/example/blob/shared/exception/PlatformNotFoundException.java
@@ -1,0 +1,7 @@
+package org.ait.project.guideline.example.blob.shared.exception;
+
+public class PlatformNotFoundException extends RuntimeException {
+  public PlatformNotFoundException() {
+    super();
+  }
+}
