+++ src/main/java/org/ait/project/guideline/example/blob/shared/exception/StorageFileNotFoundException.java
@@ -1,0 +1,7 @@
+package org.ait.project.guideline.example.blob.shared.exception;
+
+public class StorageFileNotFoundException extends RuntimeException {
+  public StorageFileNotFoundException() {
+    super();
+  }
+}
