+++ src/main/java/org/ait/project/guideline/example/config/properties/ThumbnailsConfigProperties.java
@@ -18,4 +18,6 @@ public class ThumbnailsConfigProperties {
     private String format;
 
     private int quality;
+
+    private String directory;
 }
